expose useDocumentScan hook for DocumentCameraScreen
