+++ packages/mobile-sdk-alpha/src/processing/mrz.ts
@@ -352,3 +352,20 @@ export function formatDateToYYMMDD(inputDate: string): string {
 
   throw new MrzParseError(`Invalid date format: ${inputDate}. Expected ISO format (YYYY-MM-DD) or similar.`);
 }
+
+export function checkScannedInfo(
+  passportNumber: string,
+  dateOfBirth: string,
+  dateOfExpiry: string,
+): boolean {
+  if (passportNumber.length > 9) {
+    return false;
+  }
+  if (dateOfBirth.length !== 6) {
+    return false;
+  }
+  if (dateOfExpiry.length !== 6) {
+    return false;
+  }
+  return true;
+}
